+++ src/components/badge/badge.js
@@ -4,7 +4,6 @@ import classNames from 'classnames';
 import { EuiPropTypes } from '../../utils';
 
 import { isColorDark, hexToRgb } from '../../services/color';
-import { EuiKeyboardAccessible } from '../accessibility';
 
 import { IconPropType, EuiIcon } from '../icon';
 
@@ -21,7 +20,7 @@ const colorToClassNameMap = {
 export const COLORS = Object.keys(colorToClassNameMap);
 
 const iconSideToClassNameMap = {
-  left: '',
+  left: 'euiBadge--iconLeft',
   right: 'euiBadge--iconRight',
 };
 
@@ -58,6 +57,9 @@ export const EuiBadge = ({
 
   const classes = classNames(
     'euiBadge',
+    {
+      'euiBadge-isClickable': onClick && !iconOnClick,
+    },
     iconSideToClassNameMap[iconSide],
     optionalColorClass,
     className
@@ -72,16 +74,17 @@ export const EuiBadge = ({
   if (iconType) {
     if (iconOnClick) {
       optionalIcon = (
-        <EuiKeyboardAccessible>
+        <button
+          className="euiBadge__iconButton"
+          aria-label={iconOnClickAriaLabel}
+          onClick={iconOnClick}>
           <EuiIcon
-            onClick={iconOnClick}
             type={iconType}
             size="s"
-            aria-label={iconOnClickAriaLabel}
             {...closeButtonProps}
             className={closeClassNames}
           />
-        </EuiKeyboardAccessible>
+        </button>
       );
     } else {
       optionalIcon = (
@@ -90,17 +93,32 @@ export const EuiBadge = ({
     }
   }
 
-  if (onClick) {
+  if (onClick && iconOnClick) {
+    return (
+      <span className={classes} style={optionalCustomStyles}>
+        <span className="euiBadge__content">
+          <button
+            className="euiBadge__childButton"
+            aria-label={onClickAriaLabel}
+            onClick={onClick}
+            {...rest}>
+            {children}
+          </button>
+          {optionalIcon}
+        </span>
+      </span>
+    );
+  } else if (onClick) {
     return (
       <button
-        className={classes}
-        style={optionalCustomStyles}
-        onClick={onClick}
         aria-label={onClickAriaLabel}
+        className={classes}
+        onClick={onClick}
+        style={optionalCustomStyles}
         {...rest}>
         <span className="euiBadge__content">
-          {optionalIcon}
           <span>{children}</span>
+          {optionalIcon}
         </span>
       </button>
     );
@@ -108,8 +126,8 @@ export const EuiBadge = ({
     return (
       <span className={classes} style={optionalCustomStyles} {...rest}>
         <span className="euiBadge__content">
-          {optionalIcon}
           <span className="euiBadge__text">{children}</span>
+          {optionalIcon}
         </span>
       </span>
     );