+++ src/components/form/super_select/super_select.js
@@ -195,7 +195,6 @@ export class EuiSuperSelect extends Component {
       <EuiSuperSelectControl
         options={options}
         value={valueOfSelected}
-        onChange={onChange}
         onClick={
           this.state.isPopoverOpen ? this.closePopover : this.openPopover
         }
@@ -222,6 +221,7 @@ export class EuiSuperSelect extends Component {
           style={{ width: this.state.menuWidth }}
           role="option"
           id={value}
+          aria-selected={valueOfSelected === value}
           {...optionRest}>
           {dropdownDisplay || inputDisplay}
         </EuiContextMenuItem>
@@ -251,7 +251,10 @@ export class EuiSuperSelect extends Component {
             />
           </p>
         </EuiScreenReaderOnly>
-        <div role="listbox" aria-activedescendant={valueOfSelected}>
+        <div
+          role="listbox"
+          aria-activedescendant={valueOfSelected}
+          tabIndex="0">
           {items}
         </div>
       </EuiPopover>
@@ -303,6 +306,10 @@ EuiSuperSelect.propTypes = {
    * Provides invalid styling
    */
   isInvalid: PropTypes.bool,
+  /**
+   * Provides loading indictor
+   */
+  isLoading: PropTypes.bool,
   /**
    * Make it short
    */